Added automatic Q calculation.
The code relied on a fixed variable, derivedQ, to provide the size of the step a single "click" would effect on the external controller - this was easy but slightly inefficient due to inability to account for external variations. With this update, derivedQ is calculated automatically with every change effected to the external controller, provided it is large enough to produce an effective sample. Currently, the minimum is two changes (two steps).