Added clock modifier.
Turns out (not surprisingly) that the clocking quality of crystals on Nano boards is... questionable. In order to compensate for this manufacturing drift, a new constant is added which needs to be measured on a board by board basis. Measurement can be done via oscilloscope or via frequency drift with the help of a test record (300Hz tone / 1kHz tone).